Moving relator length check to SimsTreeBase and including necessary headers.
